feat: get cabinets from database
