+++ client/src/assets/globals.scss
@@ -118,7 +118,7 @@ textarea {
   width: fit-content;
 
   &:hover {
-    background: lightgray;
+    background: lightgray !important;
   }
 }
