--- client/src/views/schedule/create-page/ClassListCard.interface.ts
@@ -1,5 +1,0 @@
-export interface ClassListCardList {
-  starred: boolean;
-  title: string;
-  points: number;
-}
