+++ client/src/views/teams/schedule/create-page/ClassListCard.interface.ts
@@ -1,0 +1,12 @@
+export interface ClassListCardList {
+  starred: boolean;
+  id: number;
+  title: string;
+  points: number;
+}
+
+export interface ClassList {
+  disabled: boolean;
+  time: string;
+  list: ClassListCardList[];
+}
